Rename testkit to kernel-testkit
