+++ core/shared/src/test/scala/cats/effect/SyncIOSpec.scala
@@ -18,7 +18,7 @@ package cats.effect
 
 import cats.kernel.laws.discipline.MonoidTests
 import cats.effect.laws.{MonadCancelTests, SyncTests}
-import cats.effect.testkit.SyncTypeGenerators
+import cats.effect.kernel.testkit.SyncTypeGenerators
 import cats.syntax.all._
 import org.scalacheck.Prop
 import Prop.forAll
