+++ kernel-testkit/shared/src/main/scala/cats/effect/kernel/testkit/FreeSyncGenerators.scala
@@ -15,17 +15,18 @@
  */
 
 package cats.effect
+package kernel
 package testkit
 
 import cats.{Eval, Monad, MonadError}
-import cats.effect.kernel.Sync
-import cats.effect.testkit.freeEval._
 import cats.free.FreeT
 
 import org.scalacheck.{Arbitrary, Cogen, Gen}
 
 import scala.concurrent.duration.FiniteDuration
 import java.util.concurrent.TimeUnit
+
+import freeEval._
 
 object FreeSyncGenerators {
