+++ kernel-testkit/shared/src/main/scala/cats/effect/kernel/testkit/Generators.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package cats.effect.testkit
+package cats.effect.kernel
+package testkit
 
 import cats.{Applicative, ApplicativeError, Eq, Monad, MonadError}
-import cats.effect.kernel._
 import cats.syntax.all._
 
 import org.scalacheck.{Arbitrary, Cogen, Gen}, Arbitrary.arbitrary
