+++ kernel-testkit/shared/src/main/scala/cats/effect/kernel/testkit/TestContext.scala
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package cats.effect.testkit
+package cats.effect.kernel
+package testkit
 
 import scala.collection.immutable.SortedSet
 import scala.concurrent.ExecutionContext
