+++ kernel-testkit/shared/src/main/scala/cats/effect/kernel/testkit/package.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.kernel
 
 import cats.data.Kleisli
