+++ laws/shared/src/test/scala/cats/effect/BaseSpec.scala
@@ -18,7 +18,7 @@ package cats.effect
 
 import cats.{Eq, FlatMap, Monad, MonadError, Show}
 import cats.data._
-import cats.effect.testkit.freeEval._
+import cats.effect.kernel.testkit.freeEval._
 import cats.laws.discipline.{eq, ExhaustiveCheck, MiniInt}, eq._
 import cats.syntax.all._
