+++ laws/shared/src/test/scala/cats/effect/laws/FreeSyncSpec.scala
@@ -18,7 +18,7 @@ package cats.effect
 package laws
 
 import cats.{Eq, Eval}
-import cats.effect.testkit.{freeEval, FreeSyncEq, FreeSyncGenerators, SyncTypeGenerators}
+import cats.effect.kernel.testkit.{freeEval, FreeSyncEq, FreeSyncGenerators, SyncTypeGenerators}
 import cats.free.FreeT
 import cats.laws.discipline.arbitrary._
 import freeEval.{syncForFreeT, FreeEitherSync}
