+++ laws/shared/src/test/scala/cats/effect/laws/StateTPureConcSpec.scala
@@ -19,7 +19,7 @@ package laws
 
 import cats.{Eq, FlatMap}
 import cats.data.StateT
-import cats.effect.testkit.{pure, PureConcGenerators}, pure._
+import cats.effect.kernel.testkit.{pure, PureConcGenerators}, pure._
 import cats.laws.discipline.{arbitrary, eq, MiniInt}, arbitrary._, eq._
 
 import org.specs2.ScalaCheck
